% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getAnnotation.R
\name{getAnnotation}
\alias{getAnnotation}
\title{Assign annotated lab + prep}
\usage{
getAnnotation(prep, lab)
}
\arguments{
\item{prep}{the prep for the RNA-seq library \code{inputParameter1}}

\item{lab}{the lab that the RNA-seq library was made from \code{inputParameter1}}
}
\value{
annotation
}
\description{
inputting prep and lab and it will return
the annotation for facet
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

