% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getGroup.R
\name{getGroup}
\alias{getGroup}
\title{Assign group for ERCC spikein}
\usage{
getGroup(fold)
}
\arguments{
\item{x}{fold change for the ERCC spikein \code{inputParameter1}}
}
\value{
group
}
\description{
inputting fold change and it will return
the group for the ERCC spikein
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

