% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getLab.R
\name{getLab}
\alias{getLab}
\title{Assign lab for the samples}
\usage{
getLab(x)
}
\arguments{
\item{x}{the sample name of the RNA-seq library \code{inputParameter1}}
}
\value{
lab name
}
\description{
inputting samplename and it will return
the lab name that the sample was prepared
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

