% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getPrep.R
\name{getPrep}
\alias{getPrep}
\title{Assign prep for the samples}
\usage{
getPrep(x)
}
\arguments{
\item{x}{the sample name of the RNA-seq library \code{inputParameter1}}
}
\value{
prep
}
\description{
inputting samplename and it will return
the prep that was used for preparing
the sequencing library
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

