% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getReplicate.R
\name{getReplicate}
\alias{getReplicate}
\title{Assign replicate for the samples}
\usage{
getReplicate(x)
}
\arguments{
\item{x}{the sample name of the RNA-seq library \code{inputParameter1}}
}
\value{
replicate
}
\description{
inputting samplename and it will return
the the replicate number that the sample was
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

