% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getTemplate.R
\name{getTemplate}
\alias{getTemplate}
\title{Assign template for the samples}
\usage{
getTemplate(x)
}
\arguments{
\item{x}{the sample name of the RNA-seq library \code{inputParameter1}}
}
\value{
template
}
\description{
inputting samplename and it will return
the template that was sequenced
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

