% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/labeling.R
\name{labeling}
\alias{labeling}
\title{Assign mix ratio for ERCC samples}
\usage{
labeling(fold)
}
\arguments{
\item{x}{designed fold change \code{inputParameter1}}
}
\value{
fold change label
}
\description{
inputting designed fold change and it will return
the the annotated label
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

